Move DeviceAuthenticator into new namespace
